feat(user): Add di module for user articles
